requirements.txt
selenium==4.15.2
pandas==2.1.3
python-dotenv==1.0.0
webdriver-manager==4.0.1